Move StandardAnswer DB layer to EE
